+++ src/Stats/sortFunction.jsx
@@ -1,0 +1,7 @@
+const sortDataByDate = (data) => {
+  const byDate = data.slice(0);
+  byDate.sort((a, b) => (b.date - a.date));
+  return byDate;
+};
+
+export default sortDataByDate;
